Boot 0 or sub-0 values from ratio stats
This change is to prevent the pipeline from crashing in the event of
errant less-than-zero predictions. Typically these occur in the
assessment set as the result of incorrect aggregations, joins, etc. We
have other methods to catch these sub-0 values and usually still want
the pipeline to run to completion in such cases.