Add unit tests for `python.comps` module (#363)
We found in https://github.com/ccao-data/model-res-avm/issues/362 that
the comps pipeline was not behaving the way we expected. This problem
was made worse by the fact that we didn't have any tests for the comps
pipeline, so we didn't have any assurances that our assumptions about
the behavior of the pipeline were true, or that they would remain
unchanged as we change the comps code in subsequent PRs.

This PR adds a simple unit test suite for the `python.comps` module as a
first step towards locking down the behavior of the comps pipeline.
These tests ensure that the Python side of the pipeline conforms to our
expectations. In a future PR, we'll want to do the same for the R side
of the pipeline.