Fix shaps and comps to work with multi-card solution (#360)
This PR addresses #358. We edit the interpret stage such that shaps and
comps are generated to be in line with our recent [multi-card
methodology
update](https://github.com/ccao-data/model-res-avm/commit/6ee2f119adbd1df016c8df71a6e25f926623b9d5).
Within multi-card PINs, we essentially compute shaps and comps with a
single-card, the card that was used for prediction in the `assess`
stage. Then, we use that card's shaps and comps for each card within the
multi-card PIN.

We also introduce another sorting mechanism which fixes an issue of
tiebreaks within multi-card PINs. We were choosing the card to predict
based on highest sqft. Sometimes the sqft was tied between two pins,
which led to a tiebreak that we didn't have a rule for. Now, we give
preference to the lowest `meta_card_num` within the PIN.

---------

Co-authored-by: Jean Cochrane <jeancochrane@users.noreply.github.com>