Configure reticulate to use ephemeral uv environments for Python dependencies (#402)
#399 updates our dependencies to allow for local Windows installations.
In the process of doing so, it tries to switch to using uv to manage
Python dependencies, but it doesn't quite get there. This PR attempts to
fix that error by configuring reticulate to use ephemeral uv
environments for Python dependencies.

As far as I can tell from my research, there are two main ways we can
use uv to manage our Python dependencies:

- **Option 1**: Install and configure Python dependencies manually via
uv, and then configure reticulate to point to the installation location
for our uv dependencies (this is what #399 attempts to do)
    - **Pro**: Affords us the full flexibility of uv
- **Con**: Requires that we update our install instructions and our CI
pipelines to reflect the extra steps required to install uv and use it
to install dependencies
        - This is the step that #399 is missing 
- **Option 2**: Configure reticulate to manage ephemeral Python
environments with uv using its [`py_require()`
function](https://pkgs.rstudio.com/reticulate/reference/py_require.html)
(this is what the current PR does)
- **Pro**: Does not require end users or CI pipelines to know about uv,
since reticulate will fully encapsulate the usage of uv for us
- **Con**: Less flexible than option 1, since it's harder to use uv
directly

Ultimately I think the tradeoffs inherent in option 2 are more
appropriate than option 1. Less flexibility means we incur more
maintenance risk down the road, if e.g. reticulate decides to make
opinionated decisions about uv usage that we disagree with, but it also
means that this package is more approachable to users who are not
comfortable managing Python environments. If you disagree with that
decision, let me know and I'd be happy to discuss it!