Remove class from complex id clusters (#406)
Based on the findings of
https://github.com/ccao-data/enterprise-intelligence/pull/323 and
conversations with res-val, we remove class as a feature in townhome
modeling.

Proof of model run with only stage 2 change:
https://github.com/ccao-data/model-res-avm/actions/runs/19117585407/job/54630593114

I also found a situation where class was selected in the ingest. I don't
see a reason why that can't be removed, however it may be useful for QC
work:
Proof of run:
https://github.com/ccao-data/model-res-avm/actions/runs/19549242919